Moved recipe.py up a level to get rid of obsolete package network.sim
